Break pageController monolith into service and controller and add unit tests
